Reduce the wavetablecache size by computing the needed cycle length to handle the number of harmonics and have enough data points for interpolation to work. Still need to investigate why the non-BL oscillator noise is coming from. (40Mb wave cache -> 7Mb).

+++ 3UTIL/util-audio/src/uk/co/modularaudio/util/audio/oscillatortable/AbstractWavetableOscillator.java
@@ -22,7 +22,7 @@ package uk.co.modularaudio.util.audio.oscillatortable;
 
 public abstract class AbstractWavetableOscillator implements Oscillator
 {
-	protected double currentPosition = 0.0;
+	protected float currentPosition = 0.0f;
 	protected WaveTableValueFetcher valueFetcher;
 	protected PulseWidthMapper pulseWidthMapper;
